Move `api/local.rs` tests into `changefeeds.rs`
