+++ lib/tests/api.rs
@@ -207,7 +207,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
 
@@ -232,7 +231,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
 
@@ -257,7 +255,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
 
@@ -282,7 +279,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
 
@@ -306,7 +302,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
 
@@ -330,7 +325,6 @@ mod api_integration {
 		}
 
 		include!("api/mod.rs");
-		include!("api/local.rs");
 		include!("api/backup.rs");
 	}
